feat: Adding support of program parameter selection on event capture
