+++ src/pages/event-capture/components/program-parameter-selection/program-parameter-selection.html
@@ -22,3 +22,34 @@
 *
 *
 -->
+<div class="animate fadeIn">
+  <div class="item-card" color="font-color" padding-top padding-bottom ion-button clear full text-center
+    text-capitalize (tap)="openOrganisationUnitTree()">
+    <img class="side-menu-icon" [src]="icons.orgUnit" />&nbsp;&nbsp;
+    <span [translate]="organisationUnitLabel"></span>
+  </div>
+  <div class="item-card" color="font-color" padding-top padding-bottom ion-button clear full text-center
+    text-capitalize (tap)="openProgramList()">
+    <img class="side-menu-icon" [src]="icons.program" />&nbsp;&nbsp;
+    <span [translate]="programLabel"></span>
+  </div>
+  <div *ngIf="programCategoryCombo && programCategoryCombo.id ">
+    <div *ngIf="isProgramDimensionApplicable">
+      <div *ngFor="let category of programCategoryCombo.categories;let categoryIndex = index">
+        <div (tap)="openDataDimensionSelection(category)" class="item-card-category" color="font-color" full
+          text-center text-wrap>
+          <div style="padding-bottom: 5px;">
+            <b>{{ category.name }}</b>
+          </div>
+          <div *ngIf="selectedDataDimension[categoryIndex]">{{ selectedDataDimension[categoryIndex].name }}</div>
+          <div *ngIf="!selectedDataDimension[categoryIndex]">
+            <span [translate]="'Touch to select option'"></span>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="item-card" padding-top padding-bottom *ngIf="!isProgramDimensionApplicable">
+      <warning [warningMessage]="programDimensionNotApplicableMessage"></warning>
+    </div>
+  </div>
+</div>
