+++ src/pages/event-capture/event-capture.html
@@ -36,67 +36,74 @@
   <ion-grid>
     <ion-row>
       <ion-col>
-        <div class="item-card" color="font-color" padding-top padding-bottom ion-button clear full text-center
-          text-capitalize (tap)="openOrganisationUnitTree()">
-          <img class="side-menu-icon" [src]="icons.orgUnit" />&nbsp;&nbsp;
-          <span *ngIf="!selectedOrgUnit || !selectedOrgUnit.name" [translate]="organisationUnitLabel"></span>
-          <span *ngIf="selectedOrgUnit && selectedOrgUnit.name"> {{ selectedOrgUnit.name }}</span>
+        <div *ngIf="!isLoading">
+          <program-parameter-selection (programParameterSelection)="onProgramParameterSelection($event)" [currentUser]="currentUser"
+            [programIdsByUserRoles]="programIdsByUserRoles" [selectedDataDimension]="selectedDataDimension"
+            [programType]="programType"></program-parameter-selection>
+          <div *ngIf="isFormReady">
+            <div class="event-storage-status">
+              <span>
+                <b>
+                  <span [translate]="'Synced'"></span>
+                </b> : {{ storageStatus.online }}
+                <b>
+                  <span [translate]="'Not synced'"></span>
+                </b> : {{storageStatus.offline }}
+
+              </span>
+              <br>
+              <span *ngIf="showEventConflictHandler" margin-left>&nbsp;&nbsp;
+                <event-conflict-handler (successEventConflictHandle)="onSuccessEventConflictHandle($event)"
+                  (successDiscovering)="onSuccessDiscoveringEvents($event)" [eventConflictHandler]="eventConflictHandler"></event-conflict-handler>
+              </span>
+
+            </div>
+
+            <div class="item-card" (click)="hideAndShowColumns()" color="font-color" padding-top padding-bottom
+              ion-button clear text-center text-capitalize icon-start>
+              <ion-icon name="ios-eye" color="icon-color"></ion-icon>
+              <span [translate]="'Columns'"></span>
+            </div>
+            <div class="item-card" (click)="goToEventRegister()" color="font-color" padding-top padding-bottom
+              ion-button clear text-center text-capitalize icon-start>
+              <ion-icon name="add" color="icon-color"></ion-icon>
+              <span [translate]="'Add'"></span>
+            </div>
+          </div>
         </div>
-        <div class="item-card" color="font-color" padding-top padding-bottom ion-button clear full text-center
-          text-capitalize (tap)="openProgramList()">
-          <img class="side-menu-icon" [src]="icons.program" />&nbsp;&nbsp;
-          <span *ngIf="!selectedProgram || !selectedProgram.name " [translate]="programLabel"></span>
-          <span *ngIf="selectedProgram && selectedProgram.name">{{ selectedProgram.name }}</span>
-        </div>
-        <div *ngIf="programCategoryCombo && programCategoryCombo.id ">
-          <div *ngIf="isProgramDimensionApplicable">
-            <div *ngFor="let category of programCategoryCombo.categories;let categoryIndex = index">
-              <div (tap)="openDataDimensionSelection(category)" class="item-card-category" color="font-color" full
-                text-center text-wrap>
-                <div style="padding-bottom: 5px;">
-                  <b>{{ category.name }}</b>
-                </div>
-                <div *ngIf="selectedDataDimension[categoryIndex]">{{ selectedDataDimension[categoryIndex].name }}</div>
-                <div *ngIf="!selectedDataDimension[categoryIndex]">
-                  <span [translate]="'Touch to select option'"></span>
+
+        <!-- <div class="item-card" color="font-color" padding-top padding-bottom ion-button clear full text-center
+            text-capitalize (tap)="openOrganisationUnitTree()">
+            <img class="side-menu-icon" [src]="icons.orgUnit" />&nbsp;&nbsp;
+            <span *ngIf="!selectedOrgUnit || !selectedOrgUnit.name" [translate]="organisationUnitLabel"></span>
+            <span *ngIf="selectedOrgUnit && selectedOrgUnit.name"> {{ selectedOrgUnit.name }}</span>
+          </div>
+          <div class="item-card" color="font-color" padding-top padding-bottom ion-button clear full text-center
+            text-capitalize (tap)="openProgramList()">
+            <img class="side-menu-icon" [src]="icons.program" />&nbsp;&nbsp;
+            <span *ngIf="!selectedProgram || !selectedProgram.name " [translate]="programLabel"></span>
+            <span *ngIf="selectedProgram && selectedProgram.name">{{ selectedProgram.name }}</span>
+          </div>
+          <div *ngIf="programCategoryCombo && programCategoryCombo.id ">
+            <div *ngIf="isProgramDimensionApplicable">
+              <div *ngFor="let category of programCategoryCombo.categories;let categoryIndex = index">
+                <div (tap)="openDataDimensionSelection(category)" class="item-card-category" color="font-color" full
+                  text-center text-wrap>
+                  <div style="padding-bottom: 5px;">
+                    <b>{{ category.name }}</b>
+                  </div>
+                  <div *ngIf="selectedDataDimension[categoryIndex]">{{ selectedDataDimension[categoryIndex].name }}</div>
+                  <div *ngIf="!selectedDataDimension[categoryIndex]">
+                    <span [translate]="'Touch to select option'"></span>
+                  </div>
                 </div>
               </div>
             </div>
-          </div>
-          <div class="item-card" padding-top padding-bottom *ngIf="!isProgramDimensionApplicable">
-            <warning [warningMessage]="translationMapper[programDimensionNotApplicableMessage]"></warning>
-          </div>
-        </div>
-        <div *ngIf="isFormReady">
-          <div class="event-storage-status">
-            <span>
-              <b>
-                <span [translate]="'Synced'"></span>
-              </b> : {{ storageStatus.online }}
-              <b>
-                <span [translate]="'Not synced'"></span>
-              </b> : {{storageStatus.offline }}
+            <div class="item-card" padding-top padding-bottom *ngIf="!isProgramDimensionApplicable">
+              <warning [warningMessage]="translationMapper[programDimensionNotApplicableMessage]"></warning>
+            </div>
+          </div> -->
 
-            </span>
-            <br>
-            <span *ngIf="showEventConflictHandler" margin-left>&nbsp;&nbsp;
-              <event-conflict-handler (successEventConflictHandle)="onSuccessEventConflictHandle($event)"
-                (successDiscovering)="onSuccessDiscoveringEvents($event)" [eventConflictHandler]="eventConflictHandler"></event-conflict-handler>
-            </span>
-
-          </div>
-
-          <div class="item-card" (click)="hideAndShowColumns()" color="font-color" padding-top padding-bottom
-            ion-button clear text-center text-capitalize icon-start>
-            <ion-icon name="ios-eye" color="icon-color"></ion-icon>
-            <span [translate]="'Columns'"></span>
-          </div>
-          <div class="item-card" (click)="goToEventRegister()" color="font-color" padding-top padding-bottom ion-button
-            clear text-center text-capitalize icon-start>
-            <ion-icon name="add" color="icon-color"></ion-icon>
-            <span [translate]="'Add'"></span>
-          </div>
-        </div>
 
       </ion-col>
     </ion-row>
