+++ src/pages/event-capture/event-capture.ts
@@ -24,12 +24,9 @@
 import { Component, OnInit } from '@angular/core';
 import { IonicPage, ModalController, NavController } from 'ionic-angular';
 import { UserProvider } from '../../providers/user/user';
-import { OrganisationUnitsProvider } from '../../providers/organisation-units/organisation-units';
-import { ProgramsProvider } from '../../providers/programs/programs';
 import { AppProvider } from '../../providers/app/app';
 import { EventCaptureFormProvider } from '../../providers/event-capture-form/event-capture-form';
 import { SettingsProvider } from '../../providers/settings/settings';
-import { AppTranslationProvider } from '../../providers/app-translation/app-translation';
 import * as _ from 'lodash';
 import { Store } from '@ngrx/store';
 import { State, getCurrentUserColorSettings } from '../../store';
@@ -52,24 +49,19 @@ export class EventCapturePage implements OnInit {
   selectedOrgUnit: any;
   selectedProgram: any;
   currentUser: any;
+  programType: string;
+  dataDimension: any;
+  selectedDataDimension: any;
   programIdsByUserRoles: Array<string>;
   isLoading: boolean;
   loadingMessage: string;
-  organisationUnitLabel: string;
-  programLabel: string;
   isFormReady: boolean = false;
-  isProgramDimensionApplicable: boolean;
-  programDimensionNotApplicableMessage: string;
-  programCategoryCombo: any;
-  selectedDataDimension: Array<any>;
-  programs: Array<any>;
   programStage: any;
   columnsToDisplay: any = {};
   icons: any = {};
   tableLayout: any;
   eventIds: Array<string>;
   currentEvents: Array<any>;
-  translationMapper: any;
   dataEntrySettings: any;
   storageStatus: any;
   showEventConflictHandler: boolean;
@@ -83,65 +75,37 @@ export class EventCapturePage implements OnInit {
     private userProvider: UserProvider,
     private modalCtrl: ModalController,
     private settingsProvider: SettingsProvider,
-    private organisationUnitsProvider: OrganisationUnitsProvider,
-    private programsProvider: ProgramsProvider,
     private appProvider: AppProvider,
     private eventCaptureFormProvider: EventCaptureFormProvider,
-    private synchronizationProvider: SynchronizationProvider,
-    private appTranslation: AppTranslationProvider
+    private synchronizationProvider: SynchronizationProvider
   ) {
+    this.programType = 'WITHOUT_REGISTRATION';
+    this.selectedDataDimension = [];
     this.colorSettings$ = this.store.select(getCurrentUserColorSettings);
     this.storageStatus = { online: 0, offline: 0 };
     this.icons = {
       accept: 'assets/icon/tick.png',
-      decline: 'assets/icon/cancel.png',
-      orgUnit: 'assets/icon/orgUnit.png',
-      program: 'assets/icon/program.png'
+      decline: 'assets/icon/cancel.png'
     };
-    this.selectedDataDimension = [];
+
     this.programIdsByUserRoles = [];
-    this.programs = [];
+
     this.isLoading = true;
     this.isFormReady = false;
-    this.isProgramDimensionApplicable = false;
-    this.translationMapper = {};
     this.showEventConflictHandler = false;
     this.eventConflictHandler = {};
     this.hasOnlineEventLoaded = false;
   }
 
   ngOnInit() {
-    this.appTranslation.getTransalations(this.getValuesToTranslate()).subscribe(
-      (data: any) => {
-        this.translationMapper = data;
-        this.loadingCurrentUserInformation();
-      },
-      error => {
-        this.loadingCurrentUserInformation();
-      }
-    );
-  }
-
-  loadingCurrentUserInformation() {
-    let key = 'Discovering current user information';
-    this.loadingMessage = this.translationMapper[key]
-      ? this.translationMapper[key]
-      : key;
+    this.loadingMessage = 'Discovering current user information';
     this.userProvider.getCurrentUser().subscribe(
       (currentUser: any) => {
         this.currentUser = currentUser;
         this.userProvider.getUserData().subscribe((userData: any) => {
           this.programIdsByUserRoles = userData.programs;
-          this.organisationUnitsProvider
-            .getLastSelectedOrganisationUnitUnit(currentUser)
-            .subscribe((lastSelectedOrgUnit: any) => {
-              if (lastSelectedOrgUnit && lastSelectedOrgUnit.id) {
-                this.selectedOrgUnit = lastSelectedOrgUnit;
-                this.loadingPrograms();
-              }
-              this.updateEventCaptureSelections();
-              this.loadingAppSetting();
-            });
+          this.loadingAppSetting();
+          this.isLoading = false;
         });
       },
       error => {
@@ -152,6 +116,32 @@ export class EventCapturePage implements OnInit {
         );
       }
     );
+  }
+
+  onProgramParameterSelection(data) {
+    const {
+      selectedOrgUnit,
+      selectedProgram,
+      dataDimension,
+      selectedDataDimension,
+      isFormReady
+    } = data;
+    this.isFormReady = isFormReady;
+    if (isFormReady) {
+      if (
+        this.selectedProgram &&
+        this.selectedProgram.id &&
+        this.selectedProgram.id !== selectedProgram.id
+      ) {
+        this.hasOnlineEventLoaded = false;
+      }
+      this.selectedDataDimension = selectedDataDimension;
+      this.selectedOrgUnit = selectedOrgUnit;
+      this.dataDimension = dataDimension;
+      this.selectedProgram = selectedProgram;
+      this.loadProgramStages(selectedProgram.id);
+      // this.loadingEvents();
+    }
   }
 
   ionViewDidEnter() {
@@ -170,219 +160,8 @@ export class EventCapturePage implements OnInit {
       });
   }
 
-  loadingPrograms() {
-    this.isLoading = true;
-    let key = 'Discovering assigned programs';
-    this.loadingMessage = this.translationMapper[key]
-      ? this.translationMapper[key]
-      : key;
-    const programType = 'WITHOUT_REGISTRATION';
-    this.programsProvider
-      .getProgramsAssignedOnOrgUnitAndUserRoles(
-        this.selectedOrgUnit.id,
-        programType,
-        this.programIdsByUserRoles,
-        this.currentUser
-      )
-      .subscribe(
-        (programs: any) => {
-          this.programs = programs;
-          this.selectedProgram = this.programsProvider.lastSelectedProgram;
-          this.updateEventCaptureSelections();
-          if (this.selectedProgram && this.selectedProgram.categoryCombo) {
-            this.updateDataSetCategoryCombo(this.selectedProgram.categoryCombo);
-            this.loadProgramStages(this.selectedProgram.id);
-          }
-          this.isLoading = false;
-          this.loadingMessage = '';
-        },
-        error => {
-          this.isLoading = false;
-          this.loadingMessage = '';
-          console.log(JSON.stringify(error));
-          this.appProvider.setNormalNotification(
-            'Failed to discover assigned programs'
-          );
-        }
-      );
-  }
-
-  updateEventCaptureSelections() {
-    if (this.organisationUnitsProvider.lastSelectedOrgUnit) {
-      this.selectedOrgUnit = this.organisationUnitsProvider.lastSelectedOrgUnit;
-      this.organisationUnitLabel = this.selectedOrgUnit.name;
-    } else {
-      this.organisationUnitLabel = 'Touch to select organisation unit';
-    }
-    if (this.selectedProgram && this.selectedProgram.name) {
-      this.programLabel = this.selectedProgram.name;
-    } else {
-      this.programLabel = 'Touch to select program';
-    }
-    this.isFormReady = this.isAllParameterSelected();
-    this.isLoading = false;
-    this.loadingMessage = '';
-    if (this.isFormReady) {
-      this.loadingEvents();
-    }
-  }
-
-  openOrganisationUnitTree() {
-    let modal = this.modalCtrl.create('OrganisationUnitSelectionPage', {
-      filterType: 'WITHOUT_REGISTRATION'
-    });
-    modal.onDidDismiss((selectedOrgUnit: any) => {
-      if (selectedOrgUnit && selectedOrgUnit.id) {
-        this.selectedOrgUnit = selectedOrgUnit;
-        this.updateEventCaptureSelections();
-        this.loadingPrograms();
-      }
-    });
-    modal.present();
-  }
-
-  openProgramList() {
-    if (this.programs && this.programs.length > 0) {
-      let modal = this.modalCtrl.create('ProgramSelection', {
-        currentProgram: this.selectedProgram,
-        programsList: this.programs
-      });
-      modal.onDidDismiss((selectedProgram: any) => {
-        if (selectedProgram && selectedProgram.id) {
-          if (
-            this.selectedProgram &&
-            this.selectedProgram.id &&
-            this.selectedProgram.id !== selectedProgram.id
-          ) {
-            this.hasOnlineEventLoaded = false;
-          }
-          this.selectedProgram = selectedProgram;
-          this.programsProvider.setLastSelectedProgram(selectedProgram);
-          this.updateEventCaptureSelections();
-          this.updateDataSetCategoryCombo(this.selectedProgram.categoryCombo);
-          this.loadProgramStages(selectedProgram.id);
-        }
-      });
-      modal.present();
-    } else {
-      this.appProvider.setNormalNotification('There are no program to select');
-    }
-  }
-
-  openDataDimensionSelection(category) {
-    if (
-      category.categoryOptions &&
-      category.categoryOptions &&
-      category.categoryOptions.length > 0
-    ) {
-      const currentIndex = this.programCategoryCombo.categories.indexOf(
-        category
-      );
-      const modal = this.modalCtrl.create('DataDimensionSelectionPage', {
-        categoryOptions: category.categoryOptions,
-        title: category.name,
-        currentSelection: this.selectedDataDimension[currentIndex]
-          ? this.selectedDataDimension[currentIndex]
-          : {}
-      });
-      modal.onDidDismiss((selectedDataDimension: any) => {
-        if (selectedDataDimension && selectedDataDimension.id) {
-          this.selectedDataDimension[currentIndex] = selectedDataDimension;
-          this.updateEventCaptureSelections();
-        }
-      });
-      modal.present();
-    } else {
-      const message =
-        'There is no option for selected category that associated with selected organisation unit';
-      this.appProvider.setNormalNotification(message);
-    }
-  }
-
-  getDataDimensions() {
-    if (this.selectedProgram && this.selectedProgram.categoryCombo) {
-      const attributeCc = this.selectedProgram.categoryCombo.id;
-      let attributeCos = '';
-      this.selectedDataDimension.forEach((dimension: any, index: any) => {
-        if (index == 0) {
-          attributeCos += dimension.id;
-        } else {
-          attributeCos += ';' + dimension.id;
-        }
-      });
-      return { attributeCc: attributeCc, attributeCos: attributeCos };
-    } else {
-      return {};
-    }
-  }
-
-  isAllParameterSelected() {
-    let isFormReady = true;
-    if (
-      this.selectedProgram &&
-      this.selectedProgram.name &&
-      this.selectedProgram.categoryCombo.name &&
-      this.selectedProgram.categoryCombo.name != 'default'
-    ) {
-      if (
-        this.selectedDataDimension &&
-        this.selectedDataDimension.length > 0 &&
-        this.programCategoryCombo &&
-        this.programCategoryCombo.categories &&
-        this.selectedDataDimension.length ==
-          this.programCategoryCombo.categories.length
-      ) {
-        let count = 0;
-        this.selectedDataDimension.forEach(() => {
-          count++;
-        });
-        if (count != this.selectedDataDimension.length) {
-          isFormReady = false;
-        }
-      } else {
-        isFormReady = false;
-      }
-    } else if (!this.selectedProgram) {
-      isFormReady = false;
-    }
-    return isFormReady;
-  }
-
-  updateDataSetCategoryCombo(categoryCombo) {
-    if (categoryCombo) {
-      let programCategoryCombo = {};
-      this.isProgramDimensionApplicable = false;
-      if (categoryCombo.name != 'default') {
-        programCategoryCombo['id'] = categoryCombo.id;
-        programCategoryCombo['name'] = categoryCombo.name;
-        let categories = this.programsProvider.getProgramCategoryComboCategories(
-          this.selectedOrgUnit.id,
-          categoryCombo.categories
-        );
-        programCategoryCombo['categories'] = categories;
-        this.isProgramDimensionApplicable = true;
-        categories.forEach((category: any) => {
-          if (
-            category.categoryOptions &&
-            category.categoryOptions.length == 0
-          ) {
-            this.isProgramDimensionApplicable = false;
-          }
-        });
-        this.programDimensionNotApplicableMessage =
-          'All of selected category disaggregation are restricted from entry in selcted organisation unit, choose a different form or contact your support desk';
-      }
-      this.selectedDataDimension = [];
-      this.programCategoryCombo = programCategoryCombo;
-      this.updateEventCaptureSelections();
-    }
-  }
-
-  loadProgramStages(programId) {
-    let key = 'Discovering program stages';
-    this.loadingMessage = this.translationMapper[key]
-      ? this.translationMapper[key]
-      : key;
+  loadProgramStages(programId, shouldLoadEvent?) {
+    this.loadingMessage = 'Discovering program stages';
     this.columnsToDisplay = {};
     this.eventCaptureFormProvider
       .getProgramStages(programId, this.currentUser)
@@ -396,7 +175,7 @@ export class EventCapturePage implements OnInit {
                 ? executionDateLabel
                 : 'Report date';
             if (this.programStage.programStageDataElements) {
-              this.programStage.programStageDataElements.forEach(
+              this.programStage.programStageDataElements.map(
                 (programStageDataElement: any) => {
                   if (
                     programStageDataElement.dataElement &&
@@ -433,7 +212,9 @@ export class EventCapturePage implements OnInit {
                 }
               );
             }
-            this.updateEventCaptureSelections();
+            if (shouldLoadEvent) {
+              this.loadingEvents();
+            }
           }
         },
         error => {
@@ -473,11 +254,8 @@ export class EventCapturePage implements OnInit {
     ) {
       this.isLoading = true;
       this.showEventConflictHandler = false;
-      let key = 'Discovering data';
-      this.loadingMessage = this.translationMapper[key]
-        ? this.translationMapper[key]
-        : key;
-      const dataDimension = this.getDataDimensions();
+      this.loadingMessage = 'Discovering data';
+      const dataDimension = this.dataDimension;
       const programId = this.selectedProgram.id;
       const programName = this.selectedProgram.name;
       const organisationUnitId = this.selectedOrgUnit.id;
@@ -629,13 +407,10 @@ export class EventCapturePage implements OnInit {
 
   renderDataAsTable() {
     this.isLoading = true;
-    let key = 'Preparing table';
     this.currentEvents = _.reverse(
       _.sortBy(_.uniqBy(this.currentEvents, 'id'), ['eventDate'])
     );
-    this.loadingMessage = this.translationMapper[key]
-      ? this.translationMapper[key]
-      : key;
+    this.loadingMessage = 'Preparing table';
     this.storageStatus.online = _.filter(
       this.currentEvents,
       event => event.syncStatus === 'synced'
@@ -652,7 +427,7 @@ export class EventCapturePage implements OnInit {
           this.eventIds = response.eventIds;
           this.isLoading = false;
         },
-        error => {
+        () => {
           this.isLoading = false;
           this.appProvider.setNormalNotification(
             'Failed to prepare table for display'
@@ -663,25 +438,14 @@ export class EventCapturePage implements OnInit {
 
   goToEventView(currentIndex) {
     let params = {
-      dataDimension: this.getDataDimensions(),
+      dataDimension: this.dataDimension,
       eventId: this.eventIds[currentIndex]
     };
     this.navCtrl.push('EventCaptureRegisterPage', params);
   }
 
   goToEventRegister() {
-    let params = { dataDimension: this.getDataDimensions() };
+    let params = { dataDimension: this.dataDimension };
     this.navCtrl.push('EventCaptureRegisterPage', params);
-  }
-
-  getValuesToTranslate() {
-    return [
-      'Discovering current user information',
-      'Discovering assigned programs',
-      'Discovering program stages',
-      'Discovering data',
-      'Preparing table for display',
-      'All of selected category disaggregation are restricted from entry in selcted organisation unit, choose a different form or contact your support desk'
-    ];
   }
 }
